refactor message service
