+++ services/messages/__init__.py
@@ -1,0 +1,1 @@
+from .messages_file import MessagesFile
